fqdn: move ExtractMsgDetails to message_handler
All callers are located there anyway.
